+++ halma.hs
@@ -3,7 +3,6 @@
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE GADTs #-}
---{-# LANGUAGE ScopedTypeVariables #-}
 
 module Main (main) where
 
@@ -52,14 +51,14 @@ startNewGame (State config _) =
 initialState :: State
 initialState = State defaultConfiguration Nothing
 
-data HalmaViewState size =
-  HalmaViewState
-  { _hvsBoard :: HalmaBoard size
-  , _hvsSelectedField :: Maybe (Int, Int)
-  , _hvsHighlightedFields :: [(Int, Int)]
-  , _hvsLastMoved :: Maybe (Int, Int)
-  , _hvsFinishedPlayers :: [Team]
-  , _hvsCompetitiveAIAllowed :: Bool
+data HalmaViewState size
+  = HalmaViewState
+  { hvsBoard :: HalmaBoard size
+  , hvsSelectedField :: Maybe (Int, Int)
+  , hvsHighlightedFields :: [(Int, Int)]
+  , hvsLastMoved :: Maybe (Int, Int)
+  , hvsFinishedPlayers :: [Team]
+  , hvsCompetitiveAIAllowed :: Bool
   } deriving (Eq, Show)
 
 data ViewState where
@@ -68,43 +67,60 @@ data ViewState where
 
 deriving instance Show ViewState
 
-data QuitType = QuitGame | QuitApp deriving (Show, Eq)
+data QuitType
+  = QuitGame
+  | QuitApp
+  deriving (Show, Eq)
 
-data AIType = Ignorant | Competitive
+data AIType
+  = Ignorant
+  | Competitive
   deriving (Eq, Show)
 
-data ViewEvent = Quit QuitType
-               | SetConfiguration SomeConfiguration
-               | NewGame
-               | FieldClick (Int, Int)
-               | EmptyClick
-               | AIMove AIType
-               deriving (Eq, Show)
+data ViewEvent
+  = Quit QuitType
+  | SetConfiguration SomeConfiguration
+  | NewGame
+  | FieldClick (Int, Int)
+  | EmptyClick
+  | AIMove AIType
+  deriving (Eq, Show)
 
 renderHalmaViewState
   :: (V b ~ V2, N b ~ Double, Renderable (Path V2 Double) b)
   => (Team -> Colour Double)
   -> HalmaViewState size
   -> QDiagram b V2 Double (Option (Last (Int, Int)))
-renderHalmaViewState teamColors (HalmaViewState board startPos highlighted lastMoved _ _) =
+renderHalmaViewState teamColors halmaViewState =
   drawBoard' (getGrid board) drawField
   where
-    drawPiece t lastMoved' =
-      let c = teamColors t
-      in circle 0.25 # fc c # lc (if lastMoved' then darken 0.2 c else darken 0.5 c) #
-        if lastMoved' then lw medium else id
+    HalmaViewState
+      { hvsBoard = board
+      , hvsSelectedField = startPos
+      , hvsHighlightedFields = highlighted
+      , hvsLastMoved = lastMoved
+      } = halmaViewState
+    drawPiece piece lastMoved' =
+      let c = teamColors (pieceTeam piece)
+      in
+        circle 0.25
+        # fc c
+        # lc (if lastMoved' then darken 0.2 c else darken 0.5 c)
+        # if lastMoved' then lw medium else id
     startField = startPos >>= flip lookupHalmaBoard board
     drawField p =
       (if Just p == startPos then lc black . lw thick else id) $
       case (lookupHalmaBoard p board, startField) of
-        (Just t, _) -> drawPiece t (Just p == lastMoved)
-        (Nothing, Just t) | p `elem` highlighted ->
-          drawPiece t False # opacity 0.5
+        (Just piece, _) -> drawPiece piece (Just p == lastMoved)
+        (Nothing, Just piece) | p `elem` highlighted ->
+          drawPiece piece False # opacity 0.5
         _ -> mempty
 
-data ButtonState a = ButtonActive a
-                   | ButtonInactive
-                   | ButtonSelected deriving (Eq, Show)
+data ButtonState a
+  = ButtonActive a
+  | ButtonInactive
+  | ButtonSelected
+  deriving (Eq, Show)
 
 button
   :: (Renderable (Path V2 Double) b, Renderable (Text Double) b)
@@ -114,19 +130,23 @@ button
 button txt buttonState = (label <> background) # value val' # padX 1.05 # padY 1.2
   where
     (label, background) = case buttonState of
-      ButtonActive _ -> ( text txt # fontSizeO 13
-                        , roundedRect 110 26 6 # fc lightgray
-                        )
-      ButtonInactive -> ( text txt # fontSizeO 13 # fc gray
-                        , roundedRect 110 26 6 # fc lightgray # lc gray
-                        )
-      ButtonSelected -> ( text txt # fontSizeO 13
-                        , roundedRect 110 26 6 # fc yellow # lc black # lw thick
-                        )
+      ButtonActive _ ->
+        ( text txt # fontSizeO 13
+        , roundedRect 110 26 6 # fc lightgray
+        )
+      ButtonInactive ->
+        ( text txt # fontSizeO 13 # fc gray
+        , roundedRect 110 26 6 # fc lightgray # lc gray
+        )
+      ButtonSelected ->
+        ( text txt # fontSizeO 13
+        , roundedRect 110 26 6 # fc yellow # lc black # lw thick
+        )
 
-    val' = case buttonState of
-      ButtonActive val -> Option (Just (Last val))
-      _ -> Option Nothing
+    val' =
+      case buttonState of
+        ButtonActive val -> Option (Just (Last val))
+        _ -> Option Nothing
 
 playerFinishedSign
   :: (Renderable (Path V2 Double) b, Renderable (Text Double) b)
@@ -154,29 +174,35 @@ renderMenu (Configuration gridSize nop) =
          )
     allButtons = case gridSize of
       SmallGrid ->
-        let nopL = case nop of
-                     TwoPlayers -> TwoPlayers
-                     ThreePlayers -> ThreePlayers
-                     _ -> error "impossible"
-        in ( [ button "Small Grid" ButtonSelected
-             , button "Large Grid" $ ButtonActive $ SomeConfiguration $ Configuration LargeGrid nopL
-             ]
-           , [ button "Two Players" $ playerButtonAction TwoPlayers
-             , button "Three Players" $ playerButtonAction ThreePlayers
-             ] ++ map (flip button ButtonInactive) ["Four Players", "Five Players", "Six Players"]
-           )
+        let
+          nopL =
+            case nop of
+              TwoPlayers -> TwoPlayers
+              ThreePlayers -> ThreePlayers
+              _ -> error "impossible"
+        in
+          ( [ button "Small Grid" ButtonSelected
+            , button "Large Grid" $ ButtonActive $ SomeConfiguration $ Configuration LargeGrid nopL
+            ]
+          , [ button "Two Players" $ playerButtonAction TwoPlayers
+            , button "Three Players" $ playerButtonAction ThreePlayers
+            ] ++ map (flip button ButtonInactive) ["Four Players", "Five Players", "Six Players"]
+          )
       LargeGrid ->
-        let nopS = case nop of
-                     TwoPlayers -> TwoPlayers
-                     _ -> ThreePlayers
-        in ( [ button "Small Grid" $ ButtonActive $ SomeConfiguration $ Configuration SmallGrid nopS
-             , button "Large Grid" ButtonSelected
-             ]
-           , map (\(numStr, nop') -> button (numStr ++ " Players") (playerButtonAction nop')) $
-             [ ("Two", TwoPlayers), ("Three", ThreePlayers), ("Four", FourPlayers)
-             , ("Five", FivePlayers), ("Six", SixPlayers)
-             ]
-           )
+        let
+          nopS =
+            case nop of
+              TwoPlayers -> TwoPlayers
+              _ -> ThreePlayers
+        in
+          ( [ button "Small Grid" $ ButtonActive $ SomeConfiguration $ Configuration SmallGrid nopS
+            , button "Large Grid" ButtonSelected
+            ]
+          , map (\(numStr, nop') -> button (numStr ++ " Players") (playerButtonAction nop')) $
+            [ ("Two", TwoPlayers), ("Three", ThreePlayers), ("Four", FourPlayers)
+            , ("Five", FivePlayers), ("Six", SixPlayers)
+            ]
+          )
 
 renderViewState
   :: (Team -> Colour Double)
@@ -189,19 +215,29 @@ renderViewState _teamColors (w,h) (MenuView config) =
       reposition    = centered (dims (r2 (w, h))) . toGtkCoords
   in reposition (menuDiagram === newGameButton)
 renderViewState teamColors (w,h) (HalmaView halmaViewState) =
-  let resize = sizedCentered (dims (r2 (w, h))) . toGtkCoords . pad 1.05
-      buttons = padY 1.3 $ quitGameButton === padY 1.3 aiButtons
-      quitGameButton = button "Quit Game" $ ButtonActive $ Quit QuitGame
-      aiButtons = button "AI Move" (aiButtonState Ignorant)
-              === if _hvsCompetitiveAIAllowed halmaViewState
-                  then button "Competitive" (aiButtonState Competitive) else mempty
-      aiButtonState aiMoveType =
-        if isJust (_hvsSelectedField halmaViewState)
-        then ButtonInactive
-        else ButtonActive $ AIMove aiMoveType
-      finishedSigns = foldl (===) mempty $
-        map (playerFinishedSign . teamColors) $ _hvsFinishedPlayers halmaViewState
-      field = resize (fmap (fmap FieldClick) <$> renderHalmaViewState teamColors halmaViewState)
+  let
+    resize = sizedCentered (dims (r2 (w, h))) . toGtkCoords . pad 1.05
+    buttons =
+      padY 1.3 $
+        quitGameButton
+        ===
+        padY 1.3 aiButtons
+    quitGameButton = button "Quit Game" $ ButtonActive $ Quit QuitGame
+    aiButtons =
+      button "AI Move" (aiButtonState Ignorant)
+      ===
+      if hvsCompetitiveAIAllowed halmaViewState then
+        button "Competitive" (aiButtonState Competitive)
+      else
+        mempty
+    aiButtonState aiMoveType =
+      if isJust (hvsSelectedField halmaViewState) then
+        ButtonInactive
+      else
+        ButtonActive $ AIMove aiMoveType
+    finishedSigns = foldl (===) mempty $
+      map (playerFinishedSign . teamColors) $ hvsFinishedPlayers halmaViewState
+    field = resize (fmap (fmap FieldClick) <$> renderHalmaViewState teamColors halmaViewState)
   in toGtkCoords (buttons === finishedSigns) `atop` field
 
 external :: Managed (View ViewState, Controller ViewEvent)
@@ -214,32 +250,33 @@ external = managed $ \f -> do
   viewState <- newEmptyMVar
   (veOutput, veInput) <- spawn (bounded 1)
 
-  let figure winSize =
-        fmap (maybe mempty (renderViewState defaultTeamColours winSize))
-             (tryReadMVar viewState)
-      resizedFigure = do
-        drawWin <- widgetGetDrawWindow canvas
-        (w, h) <- drawableGetSize drawWin
-        figure (fromIntegral w, fromIntegral h)
-      renderFigure = tryEvent $ do
-        win <- eventWindow
-        liftIO $ putStr "Render time: "
-        liftIO $ timeIt $ resizedFigure >>= renderToGtk win
-      updateViewState vs = do
-        _ <- tryTakeMVar viewState
-        putMVar viewState vs
-        widgetQueueDraw canvas -- send redraw request to canvas
-      handleClick = tryEvent $ do
-        _click <- eventClick
-        (x,y) <- eventCoordinates
-        fig <- liftIO resizedFigure
-        let result = runQuery (query fig) (P (r2 (x, y)))
-            event = maybe EmptyClick getLast $ getOption result
-        liftIO $ print event
-        void $ liftIO $ atomically $ send veOutput event
-      handleDestroy = do
-        _ <- atomically $ send veOutput $ Quit QuitApp
-        mainQuit
+  let
+    figure winSize =
+      fmap (maybe mempty (renderViewState defaultTeamColours winSize))
+           (tryReadMVar viewState)
+    resizedFigure = do
+      drawWin <- widgetGetDrawWindow canvas
+      (w, h) <- drawableGetSize drawWin
+      figure (fromIntegral w, fromIntegral h)
+    renderFigure = tryEvent $ do
+      win <- eventWindow
+      liftIO $ putStr "Render time: "
+      liftIO $ timeIt $ resizedFigure >>= renderToGtk win
+    updateViewState vs = do
+      _ <- tryTakeMVar viewState
+      putMVar viewState vs
+      widgetQueueDraw canvas -- send redraw request to canvas
+    handleClick = tryEvent $ do
+      _click <- eventClick
+      (x,y) <- eventCoordinates
+      fig <- liftIO resizedFigure
+      let result = runQuery (query fig) (P (r2 (x, y)))
+          event = maybe EmptyClick getLast $ getOption result
+      liftIO $ print event
+      void $ liftIO $ atomically $ send veOutput event
+    handleDestroy = do
+      _ <- atomically $ send veOutput $ Quit QuitApp
+      mainQuit
 
   _ <- canvas `on` sizeRequest $ return (Requisition 650 450)
   _ <- canvas `on` exposeEvent $ renderFigure
@@ -255,17 +292,31 @@ gameLoop
   :: Configuration size
   -> HalmaState size
   -> Pipe ViewEvent (HalmaViewState size) (MS.State State) QuitType
-gameLoop config (HalmaState ruleOptions board turnCounter lastMoved) = noSelectionLoop
+gameLoop config halmaState = noSelectionLoop
   where
+    HalmaState
+      { hsRuleOptions = ruleOptions
+      , hsBoard = board
+      , hsTurnCounter = turnCounter
+      , hsLastMoved = lastMoved
+      } = halmaState
     team = currentPlayer turnCounter
     finishedPlayers = filter (hasFinished board) (tcPlayers turnCounter)
+    competitiveAIAllowed = length (tcPlayers turnCounter) == 2
     noSelectionLoop = do
-      yield $ HalmaViewState board Nothing [] lastMoved finishedPlayers
-                             (length (tcPlayers turnCounter)==2)
+      yield $
+        HalmaViewState
+        { hvsBoard = board
+        , hvsSelectedField = Nothing
+        , hvsHighlightedFields = []
+        , hvsLastMoved = lastMoved
+        , hvsFinishedPlayers = finishedPlayers
+        , hvsCompetitiveAIAllowed = competitiveAIAllowed
+        }
       event <- await
       case event of
         EmptyClick -> noSelectionLoop
-        FieldClick p | lookupHalmaBoard p board == Just team ->
+        FieldClick p | (pieceTeam <$> lookupHalmaBoard p board) == Just team ->
           selectionLoop p
         FieldClick _ -> noSelectionLoop
         AIMove Ignorant -> performMove $
@@ -277,20 +328,34 @@ gameLoop config (HalmaState ruleOptions board turnCounter lastMoved) = noSelecti
         _ -> return QuitApp
     selectionLoop startPos = do
       let possible = possibleMoves ruleOptions board startPos
-      yield $ HalmaViewState board (Just startPos) possible Nothing finishedPlayers
-                             (length (tcPlayers turnCounter)==2)
+      yield $
+        HalmaViewState
+        { hvsBoard = board
+        , hvsSelectedField = Just startPos
+        , hvsHighlightedFields = possible
+        , hvsLastMoved = Nothing
+        , hvsFinishedPlayers = finishedPlayers
+        , hvsCompetitiveAIAllowed = competitiveAIAllowed
+        }
       event <- await
       case event of
         EmptyClick -> noSelectionLoop
         FieldClick p | p `elem` possible -> performMove (startPos, p)
-        FieldClick p | lookupHalmaBoard p board == Just team ->
+        FieldClick p | (pieceTeam <$> lookupHalmaBoard p board) == Just team ->
           selectionLoop p
         FieldClick _ -> noSelectionLoop
         Quit quitType -> return quitType
         _ -> return QuitApp
     performMove (source, destination) = do
-      let Right board' = movePiece source destination board
-          halmaState' = HalmaState ruleOptions board' (nextTurn turnCounter) (Just destination)
+      let
+        Right board' = movePiece source destination board
+        halmaState' =
+          HalmaState
+          { hsRuleOptions = ruleOptions
+          , hsBoard = board'
+          , hsTurnCounter = nextTurn turnCounter
+          , hsLastMoved = Just destination
+          }
       MS.put $ State config (Just halmaState')
       gameLoop config halmaState'